refactor(Timeline): Split out hook
Obeys single responsibility
