+++ packages/react-component-library/src/components/Timeline/hooks/useTimelineFrame.ts
@@ -1,47 +1,17 @@
-import { useContext, useEffect, useState } from 'react'
-import { isNil } from 'lodash'
+import { useContext } from 'react'
 
-import { TIMELINE_ACTIONS, TimelineScaleOption } from '../context/types'
-import { TimelineContext } from '../context'
 import { initialiseScaleOptions } from '../context/timeline_scales'
+import { TimelineContext } from '../context'
+import { TIMELINE_ACTIONS, TimelineScaleOption } from '../context/types'
 
-export function useTimelineScale() {
+export function useTimelineFrame(): {
+  moveNext: () => void
+  movePrevious: () => void
+} {
   const {
     dispatch,
-    state: {
-      currentScaleIndex,
-      currentScaleOption,
-      options,
-      scaleOptions,
-      width,
-    },
+    state: { currentScaleOption, options, width },
   } = useContext(TimelineContext)
-  const [timelineScaleOptions, setTimelineScaleOptions] = useState(scaleOptions)
-  const canZoomIn = isNil(currentScaleIndex) || currentScaleIndex > 0
-  const canZoomOut = currentScaleIndex < timelineScaleOptions.length - 1
-
-  useEffect(() => {
-    const newScaleOptions = initialiseScaleOptions(options, width)
-    setTimelineScaleOptions(newScaleOptions)
-  }, [width])
-
-  useEffect(() => {
-    setTimelineScaleOptions(scaleOptions)
-  }, [scaleOptions])
-
-  function zoomIn() {
-    dispatch({
-      scaleIndex: currentScaleIndex - 1,
-      type: TIMELINE_ACTIONS.SCALE,
-    })
-  }
-
-  function zoomOut() {
-    dispatch({
-      scaleIndex: currentScaleIndex + 1,
-      type: TIMELINE_ACTIONS.SCALE,
-    })
-  }
 
   function move(
     type: typeof TIMELINE_ACTIONS.GET_NEXT | typeof TIMELINE_ACTIONS.GET_PREV
@@ -90,11 +60,7 @@ export function useTimelineScale() {
   }
 
   return {
-    canZoomIn,
-    canZoomOut,
     moveNext,
     movePrevious,
-    zoomIn,
-    zoomOut,
   }
 }
